+++ jest.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  testEnvironment: 'node',
+  roots: ['<rootDir>/tests'],
+  testMatch: ['**/*.test.js', '**/*.spec.js'],
+  collectCoverageFrom: [
+    'scripts/**/*.sh',
+    '!**/node_modules/**',
+    '!**/vendor/**',
+  ],
+  coverageDirectory: 'coverage',
+  coverageReporters: ['text', 'lcov', 'html'],
+  verbose: true,
+  testTimeout: 30000,
+};